requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
pyjwt
bcrypt
